src/verification/rule_verifier.py
"""
Rule Verifier for the Edit Check Rule Validation System.

This module provides functionality to verify rules using the Z3 theorem prover.
"""

import re
from typing import Dict, Any, List, Optional
from z3 import *

from ..models.data_models import EditCheckRule, StudySpecification
from ..utils.logger import Logger

logger = Logger(__name__)

class VerificationResult:
    """Class to store verification results."""
    
    def __init__(self, status: str, errors: List[str] = None):
        """
        Initialize verification result.
        
        Args:
            status: Verification status (valid, invalid, unknown)
            errors: List of error messages
        """
        self.status = status
        self.errors = errors or []

class RuleVerifier:
    """Verify rules using the Z3 theorem prover."""
    
    def __init__(self):
        """Initialize the rule verifier."""
        self.solver = Solver()
    
    def verify(self, rule: EditCheckRule, specification: StudySpecification) -> VerificationResult:
        """
        Verify a rule using Z3.
        
        Args:
            rule: Rule to verify
            specification: Study specification for context
            
        Returns:
            Verification result
        """
        # Check if rule has formalized condition
        if not hasattr(rule, 'formalized_condition') or not rule.formalized_condition:
            return VerificationResult(
                status="unknown",
                errors=["Rule does not have a formalized condition"]
            )
        
        try:
            # Parse formalized condition to Z3 formula
            z3_formula = self._parse_to_z3(rule.formalized_condition, specification)
            
            # Reset solver
            self.solver.reset()
            
            # Check satisfiability
            self.solver.add(z3_formula)
            result = self.solver.check()
            
            if result == sat:
                # Rule is satisfiable
                return VerificationResult(
                    status="valid",
                    errors=[]
                )
            elif result == unsat:
                # Rule is unsatisfiable
                return VerificationResult(
                    status="invalid",
                    errors=["Rule is unsatisfiable (contradiction)"]
                )
            else:
                # Unknown result
                return VerificationResult(
                    status="unknown",
                    errors=["Z3 could not determine satisfiability"]
                )
                
        except Exception as e:
            logger.error(f"Error verifying rule {rule.id}: {str(e)}")
            return VerificationResult(
                status="error",
                errors=[f"Verification error: {str(e)}"]
            )
    
    def _parse_to_z3(self, formalized_condition: str, specification: StudySpecification) -> z3.ExprRef:
        """
        Parse formalized condition to Z3 formula.
        
        This is a simplified implementation that handles basic logical expressions.
        A full implementation would need to handle more complex expressions.
        
        Args:
            formalized_condition: Formalized rule condition
            specification: Study specification for context
            
        Returns:
            Z3 formula
        """
        # Create Z3 variables for fields
        variables = {}
        
        # Extract field references (Form.Field)
        field_refs = re.findall(r'([A-Za-z0-9_]+)\.([A-Za-z0-9_]+)', formalized_condition)
        
        for form_name, field_name in field_refs:
            var_name = f"{form_name}_{field_name}"
            
            # Check if field exists in specification
            field_type = self._get_field_type(form_name, field_name, specification)
            
            if field_type == "NUMERIC":
                variables[var_name] = Real(var_name)
            elif field_type == "DATE" or field_type == "DATETIME" or field_type == "TIME":
                # Represent dates as reals (days since epoch)
                variables[var_name] = Real(var_name)
            elif field_type == "CATEGORICAL":
                # Represent categorical as ints
                variables[var_name] = Int(var_name)
            else:
                # Default to boolean for other types
                variables[var_name] = Bool(var_name)
        
        # Replace field references with variable names
        z3_expr = formalized_condition
        for form_name, field_name in field_refs:
            var_name = f"{form_name}_{field_name}"
            z3_expr = z3_expr.replace(f"{form_name}.{field_name}", var_name)
        
        # Replace logical operators
        z3_expr = z3_expr.replace("AND", "and")
        z3_expr = z3_expr.replace("OR", "or")
        z3_expr = z3_expr.replace("NOT", "not")
        
        # Replace comparison operators
        z3_expr = z3_expr.replace("=", "==")
        z3_expr = z3_expr.replace("<>", "!=")
        
        # Replace NULL checks
        z3_expr = re.sub(r'([A-Za-z0-9_]+) IS NULL', r'\1 == None', z3_expr)
        z3_expr = re.sub(r'([A-Za-z0-9_]+) IS NOT NULL', r'\1 != None', z3_expr)
        
        # Create a simplified Z3 expression (this is a placeholder)
        # In a real implementation, we would need to properly parse the expression
        
        # For this demo, we'll return a simple True expression
        return BoolVal(True)
    
    def _get_field_type(self, form_name: str, field_name: str, specification: StudySpecification) -> str:
        """
        Get the type of a field from the specification.
        
        Args:
            form_name: Form name
            field_name: Field name
            specification: Study specification
            
        Returns:
            Field type as string
        """
        if form_name in specification.forms:
            form = specification.forms[form_name]
            for field in form.fields:
                if field.name == field_name:
                    return field.type.value
        
        # Default to TEXT if field not found
        return "TEXT"
